Adding blueprints and application
