+++ .env
@@ -1,0 +1,14 @@
+# .env
+
+# Database Configuration
+DEV_DATABASE_URI=mysql+mysqldb://cbase_dev:cbase_dev_pwd@localhost/cbase_dev_db
+TEST_DATABASE_URI=mysql+mysqldb://cbase_test:cbase_test_pwd@localhost/cbase_test_db
+
+# Flask Configuration
+SECRET_KEY=4a7ad68f727fb8743fc5751b0a345cb0
+
+# Admin Login Information
+ADMIN_EMAIL=cbase_admin@gmail.com
+ADMIN_PASSWORD=cbase_2023
+ADMIN_ID=12346879
+ADMIN_NAME=ClinicBase
