+++ .gitignore
@@ -1,4 +1,3 @@
 __pycache__
-.env
 ~
 .swp
