+++ clinic_app.py
@@ -1,45 +1,17 @@
 #!/usr/bin/python3
-""" Dummy content for application """
+""" Flask Application """
 
-from flask import Flask
-from flask_bcrypt import Bcrypt
-from flask_login import LoginManager
-from config import Config
+from dashboards import create_app
 from models import storage
 
-bcrypt = Bcrypt()
-login_manager = LoginManager()
-login_manager.login_view = 'auth.login'
-login_manager.login_message_category = 'info'
-
-
-def create_app(config_class=Config):
-    app = Flask(__name__)
-    app.config.from_object(Config)
-
-    bcrypt.init_app(app)
-    login_manager.init_app(app)
-
-    """
-    from flaskblog.users.routes import users
-    from flaskblog.posts.routes import posts
-    from flaskblog.main.routes import main
-    from flaskblog.errors.handlers import errors
-    app.register_blueprint(users)
-    app.register_blueprint(posts)
-    app.register_blueprint(main)
-    app.register_blueprint(errors)
-    """
-
-    return app
-
-
 app = create_app()
+
 
 @app.teardown_appcontext
 def close_db(error):
     """ Remove the current SQLAlchemy Session """
     storage.close()
+
 
 if __name__ == '__main__':
     app.run(host='0.0.0.0', port=5000)
