+++ dashboards/landing_page/__init__.py
@@ -1,0 +1,8 @@
+#!/usr/bin/python3
+""" Blueprint for main """
+
+from flask import Blueprint
+
+bp_main = Blueprint('main', __name__)
+
+from dashboards.landing_page import routes
